requirements.txt
asgiref==3.8.1
Django==5.1.3
pytz==2024.1
sqlparse==0.5.0
debugpy==1.8.6
ruff==0.9.6
colorlog==6.8.2

djangorestframework==3.14.0
django-allauth==0.58.2
django-cors-headers==4.3.1
requests==2.31.0
Pillow==10.1.0